# docs/swagger.yaml
definitions:
  common.JSONResult:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
host: https://dev.back.recruitment2023.hustunique.com/
info:
  contact:
    email: wwbstar07@gmail.com
  description: This is API doc of UniqueStudio Recruitment. For more API information,
    please see https://app.apifox.com/project/2985744
  title: UniqueStudio Recruitment API
  version: "0.1"
paths:
...
  /sms:
    post:
      consumes:
      - application/json
      description: Send code to member(todo)
      operationId: send_code
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: Send code
      tags:
      - Sms
...
swagger: "2.0"